common/rc: introduce minimal fs size check
_scratch_mkfs_sized() create a file system with specified size limit. It
can, however, too small for certain kind of devices. For example, a file
system on a zoned device may require some number of zones to make a file
system.

This commit introduces MIN_FSSIZE, which specify the minimum size of the
possible file system. We can set this variable e.g. $ZONE_SIZE *
$MIN_ZONE_COUNT.

Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>